move function to UserAccess model
