+++ src/api.v2/controllers/accessController.js
@@ -1,4 +1,4 @@
-const Experiment = require('../model/Experiment');
+const UserAccess = require('../model/UserAccess');
 const getLogger = require('../../utils/getLogger');
 
 const logger = getLogger('[AccessController] - ');
@@ -7,7 +7,7 @@ const getExperimentUsers = async (req, res) => {
   const { experimentId } = req.params;
 
   logger.log(`Fetching users for experiment ${experimentId}`);
-  const data = await new Experiment().getUsers(experimentId);
+  const data = await new UserAccess().getExperimentUsers(experimentId);
 
   logger.log(`Users fetched for experiment ${experimentId}`);
